Added linting to provider (#262)
* updated MAKEFILE & added golangci-lint

* Fixes linting checks/errors

+++ github/resource_github_organization_webhook_test.go
@@ -159,7 +159,7 @@ func testAccCheckGithubOrganizationWebhookDestroy(s *terraform.State) error {
 
 		gotHook, resp, err := conn.Organizations.GetHook(context.TODO(), orgName, id)
 		if err == nil {
-			if gotHook != nil && *gotHook.ID == int64(id) {
+			if gotHook != nil && *gotHook.ID == id {
 				return fmt.Errorf("Webhook still exists")
 			}
 		}
